requirements.txt
openpyxl
fuzzywuzzy
python-magic
xlrd
streamlit